+++ tests/QueryBuilder/DSL/AggregationTest.php
@@ -34,7 +34,7 @@ class AggregationTest extends AbstractDSLTest
 
         $this->_assertImplemented($aggregationDSL, 'avg', Aggregation\Avg::class, ['name']);
         $this->_assertImplemented($aggregationDSL, 'cardinality', Aggregation\Cardinality::class, ['name']);
-        $this->_assertImplemented($aggregationDSL, 'date_histogram', Aggregation\DateHistogram::class, ['name', 'field', 1]);
+        $this->_assertImplemented($aggregationDSL, 'date_histogram', Aggregation\DateHistogram::class, ['name', 'field']);
         $this->_assertImplemented($aggregationDSL, 'date_range', Aggregation\DateRange::class, ['name']);
         $this->_assertImplemented($aggregationDSL, 'extended_stats', Aggregation\ExtendedStats::class, ['name']);
         $this->_assertImplemented($aggregationDSL, 'filter', Aggregation\Filter::class, ['name', new Exists('field')]);